requirements.txt
#Requirements

Flask
flask-cors
gunicorn
google-cloud-firestore
textblob
Pillow